requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
Deprecated==1.2.14
Django==4.2.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
idna==3.4
jwcrypto==1.5.0
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.3
python-jwt==4.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.5
wrapt==1.15.0
